// TenderManagementSystem.API/ClientApp/src/app/tender-management-api.ts
export class TenderDTO implements ITenderDTO {
    id?: number;
    name?: string | undefined;
    contractNumber?: string | undefined;
    releaseDate?: Date | undefined;
    closingDate?: Date | undefined;
    description?: string | undefined;

    constructor(data?: ITenderDTO) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
            this.contractNumber = _data["contractNumber"];
            this.releaseDate = _data["releaseDate"] ? new Date(_data["releaseDate"].toString()) : <any>undefined;
            this.closingDate = _data["closingDate"] ? new Date(_data["closingDate"].toString()) : <any>undefined;
            this.description = _data["description"];
        }
    }

    static fromJS(data: any): TenderDTO {
        data = typeof data === 'object' ? data : {};
        let result = new TenderDTO();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        data["contractNumber"] = this.contractNumber;
        data["releaseDate"] = this.releaseDate ? this.releaseDate.toISOString() : <any>undefined;
        data["closingDate"] = this.closingDate ? this.closingDate.toISOString() : <any>undefined;
        data["description"] = this.description;
        return data; 
    }
}

export interface ITenderDTO {
    id?: number;
    name?: string | undefined;
    contractNumber?: string | undefined;
    releaseDate?: Date | undefined;
    closingDate?: Date | undefined;
    description?: string | undefined;
}

export class AddCommand implements IAddCommand {
    name?: string | undefined;
    contractNumber?: string | undefined;
    releaseDate?: string | undefined;
    closingDate?: string | undefined;
    description?: string | undefined;

    constructor(data?: IAddCommand) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.name = _data["name"];
            this.contractNumber = _data["contractNumber"];
            this.releaseDate = _data["releaseDate"];
            this.closingDate = _data["closingDate"];
            this.description = _data["description"];
        }
    }

    static fromJS(data: any): AddCommand {
        data = typeof data === 'object' ? data : {};
        let result = new AddCommand();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["name"] = this.name;
        data["contractNumber"] = this.contractNumber;
        data["releaseDate"] = this.releaseDate;
        data["closingDate"] = this.closingDate;
        data["description"] = this.description;
        return data; 
    }
}

export interface IAddCommand {
    name?: string | undefined;
    contractNumber?: string | undefined;
    releaseDate?: string | undefined;
    closingDate?: string | undefined;
    description?: string | undefined;
}

export class UpdateCommand implements IUpdateCommand {
    id?: number;
    name?: string | undefined;
    contractNumber?: string | undefined;
    releaseDate?: string | undefined;
    closingDate?: string | undefined;
    description?: string | undefined;

    constructor(data?: IUpdateCommand) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
            this.contractNumber = _data["contractNumber"];
            this.releaseDate = _data["releaseDate"];
            this.closingDate = _data["closingDate"];
            this.description = _data["description"];
        }
    }

    static fromJS(data: any): UpdateCommand {
        data = typeof data === 'object' ? data : {};
        let result = new UpdateCommand();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        data["contractNumber"] = this.contractNumber;
        data["releaseDate"] = this.releaseDate;
        data["closingDate"] = this.closingDate;
        data["description"] = this.description;
        return data; 
    }
}

export interface IUpdateCommand {
    id?: number;
    name?: string | undefined;
    contractNumber?: string | undefined;
    releaseDate?: string | undefined;
    closingDate?: string | undefined;
    description?: string | undefined;
}